:recycle:(refactor): created componente cards
